Cumulative Events

An S4 class to store the result of a tempo plot.

Slots

lower

A numeric vector giving the lower boundaries of the credibility interval expressed in rata die.

upper

A numeric vector giving the upper boundaries of the credibility interval expressed in rata die.

level

A length-one numeric vector giving the confidence level.

gauss

A logical scalar indicating if the Gaussian approximation of the credible interval was used.

counts

A logical scalar.

events

An integer scalar giving the number of events included in the tempo plot.

hash

A character string giving the 32-byte MD5 hash of the original data file.

Note

This class inherits from aion::TimeSeries.

Coerce

In the code snippets below, x is a CumulativeEvents object.

as.data.frame(x)

Coerces to a data.frame.

See also

Author

N. Frerebeau