Coerce

Coerce

as_mcmc(from, ...)

as_phases(from, ...)

# S4 method for matrix
as_mcmc(from, BP = FALSE, iteration = NULL)

# S4 method for matrix
as_phases(from, start, end, BP = FALSE, iteration = NULL)

# S4 method for data.frame
as_mcmc(from, BP = FALSE, iteration = NULL)

# S4 method for data.frame
as_phases(from, start, end, BP = FALSE, iteration = NULL)

Arguments

from
from An object to be coerced.

...
Currently not used.

BP
A logical scalar: should the data be converted from BP to BC/AD? This should not be TRUE unless you change the default settings in 'OxCal' or 'ChronoModel'.

iteration
An integer specifying the index of the iteration column to be removed.

start
An integer vector specifying the index of the columns corresponding to the beginning of the phases. If missing, every other column is used starting from the first column (after deleting the iteration column, if any).

end
An integer vector specifying the index of the columns corresponding to the end of the phases. If missing, every other column is used starting from the second column (after deleting the iteration column, if any).

Value

An MCMC or a PhasesMCMC object.

Note

If start or end is missing, the default value of corresponds to a CSV file exported from 'ChronoModel'.

See also

Author

A. Philippe, M.-A. Vibet, N. Frerebeau

Examples

## Coerce to MCMC
events <- as_mcmc(Events, iteration = 1)

## Coerce to Phases
phases <- as_phases(Phases, iteration = 1)