Hiatus Between Two Dates

Tests for the existence of a hiatus between two parameters.

hiatus(x, y, ...)

# S4 method for numeric,numeric
hiatus(x, y, level = 0.95)

# S4 method for EventsMCMC,missing
hiatus(x, level = 0.95)

# S4 method for PhasesMCMC,missing
hiatus(x, level = 0.95)

Arguments

x, y

A numeric vector. If y is missing, x must be an PhasesMCMC or an EventsMCMC object.

...

Currently not used.

level

A length-one numeric vector giving the confidence level.

Value

The endpoints of the hiatus between successive events/phases (at a given level).

Details

Finds if a gap exists between two dates and returns the longest interval that satisfies \(P(x < HiatusInf < HiatusSup < y | M) = level\)

The hiatus between two successive phases is the longest interval that satisfies \(P(Phase1Max < IntervalInf < IntervalSup < Phase2Min | M) = level\) (this assumes that the phases are in temporal order constraint).

Methods (by class)

hiatus(x = numeric, y = numeric): Returns a length-three numeric vector (terminal times and hiatus duration, if any).

hiatus(x = EventsMCMC, y = missing): Returns a TimeRange object.

hiatus(x = PhasesMCMC, y = missing): Returns a TimeRange object.

See also

Author

A. Philippe, M.-A. Vibet, N. Frerebeau

Examples

## Coerce to MCMC
eve <- as_events(mcmc_events, calendar = CE(), iteration = 1)
eve <- eve[1:10000, ]

## Test for anteriority
older(eve)
#>        E1    E2     E3     E4
#> E1 0.0000 0e+00 0.4925 0.0000
#> E2 1.0000 0e+00 1.0000 0.9996
#> E3 0.5075 0e+00 0.0000 0.0000
#> E4 1.0000 4e-04 1.0000 0.0000

## Test for hiatus
hia <- hiatus(eve)
as.data.frame(hia)
#>           start        end  duration
#> E2-E1 -1601.050 -1059.8011 542.24640
#> E4-E1 -1062.599 -1045.3355  18.26092
#> E2-E3 -1603.777  -803.1072 801.66934
#> E4-E3 -1067.124  -793.8029 274.31905
#> E2-E4 -1599.472 -1408.9400 191.53059