Read MCMC Output

Read MCMC Output

read_oxcal(file, ...)

read_bcal(file, ...)

read_chronomodel(file, ...)

# S4 method for character
read_oxcal(file, BP = FALSE)

# S4 method for character
read_bcal(file, bin_width = 1)

# S4 method for character
read_chronomodel(file, BP = FALSE, sep = ",", dec = ".")

Arguments

file
the name of the file which the data are to be read from. Each row of the table appears as one line of the file. If it does not contain an absolute path, the file name is relative to the current working directory, getwd(). Tilde-expansion is performed where supported. This can be a compressed file (see file).

Alternatively, file can be a readable text-mode connection (which will be opened for reading if necessary, and if so closed (and hence destroyed) at the end of the function call). (If stdin() is used, the prompts for lines may be somewhat confusing. Terminate input with a blank line or an EOF signal, Ctrl-D on Unix and Ctrl-Z on Windows. Any pushback on stdin() will be cleared before return.)

file can also be a complete URL. (For the supported URL schemes, see the ‘URLs’ section of the help for url.)

...
Further arguments to be passed to read.table.

BP
A logical scalar: should the data be converted from BP to BC/AD? This should not be TRUE unless you change the default settings in 'OxCal' or 'ChronoModel'.

bin_width
The bin width specified for the 'BCal' calibration. Defaults to the 'BCal' default of 1.

sep
the field separator character. Values on each line of the file are separated by this character. If sep = "" (the default for read.table) the separator is ‘white space’, that is one or more spaces, tabs, newlines or carriage returns.

dec
the character used in the file for decimal points.

Value

An MCMC object.

References

Bronk Ramsey, C. (2009). Bayesian Analysis of Radiocarbon Dates. Radiocarbon, 51(1), 337-360. doi: 10.1017/S0033822200033865 .

Buck C. E., Christen J. A. & James G. N. (1999). BCal: an on-line Bayesian radiocarbon calibration tool. Internet Archaeology, 7. doi: 10.11141/ia.7.1 .

Lanos, Ph., Philippe, A., Lanos, H. & Dufresne, Ph. (2015). Chronomodel: Chronological Modeling of Archaeological Data using Bayesian Statistics. Version 2.0.18. URL: https://www.chronomodel.fr.

See also

Author

T. S. Dye, N. Frerebeau

Examples

if (FALSE) {
## Import ChronoModel Output
path_zip <- system.file("extdata/chronomodel.zip", package = "ArchaeoPhases")
path_csv <- utils::unzip(path_zip, exdir = tempdir())
events <- read_chronomodel(path_csv[[1]])
phases <- read_chronomodel(path_csv[[2]])

## Import OxCal Output
path_zip <- system.file("extdata/oxcal.zip", package = "ArchaeoPhases")
path_csv <- utils::unzip(path_zip, exdir = tempdir())
oxcal <- read_chronomodel(path_csv)

is_original(oxcal, path_csv)
}