Read BCal Output

Reads MCMC output.

Usage

read_bcal(file, ...)

# S4 method for class 'character'
read_bcal(file, bin_width = 1, calendar = BP())

Arguments

file

A character string giving the name of the CSV file which the data are to be read from.

...

Currently not used.

bin_width

The bin width specified for the BCal calibration. Defaults to the BCal default of 1.

calendar

A aion::TimeScale object specifying the calendar (see calendar()). It should be BP() unless you change the default settings in 'BCal'.

Value

An EventsMCMC object.

References

Buck C. E., Christen J. A. & James G. N. (1999). BCal: an on-line Bayesian radiocarbon calibration tool. Internet Archaeology, 7. doi:10.11141/ia.7.1 .

See also

Author

T. S. Dye, N. Frerebeau

Examples

if (requireNamespace("ArchaeoData", quietly = TRUE)) {
  ## Import BCal Output
  path_output <- system.file("bcal/fishpond.csv", package = "ArchaeoData")
  (bcal <- read_bcal(path_output))
}
#> <EventsMCMC>
#> - Number of events: 10
#> - Number of MCMC samples: 55964