ArchaeoPhases: Post-Processing of Markov Chain Monte Carlo Simulations for Chronological Modelling

Statistical analysis of archaeological dates and groups of dates. This package allows to post-process Markov Chain Monte Carlo (MCMC) simulations from 'ChronoModel' https://chronomodel.com/, 'Oxcal' https://c14.arch.ox.ac.uk/oxcal.html or 'BCal' https://bcal.shef.ac.uk/. It provides functions for the study of rhythms of the long term from the posterior distribution of a series of dates (tempo and activity plot). It also allows the estimation and visualization of time ranges from the posterior distribution of groups of dates (e.g. duration, transition and hiatus between successive phases) as described in Philippe and Vibet (2020) doi:10.18637/jss.v093.c01 .

Details

Package:	ArchaeoPhases

Type:	Package

Version:	2.0.0

License:	GPL-3

Zenodo:	doi:10.5281/zenodo.8087121

JSS:	doi:10.18637/jss.v093.c01

Package options

ArchaeoPhases uses the following options() to configure behaviour:

ArchaeoPhases.calendar: a aion::TimeScale object (default calendar for printing).

ArchaeoPhases.grid: a numeric value specifying the number of equally spaced points at which densities are to be estimated (defaults to \(512\)). Should be a power of \(2\).

ArchaeoPhases.precision: an integer indicating the number of decimal places (defaults to \(0\)).

ArchaeoPhases.progress: a logical scalar specifying if progress bars should be displayed (defaults to interactive()).

See also

Author

Full list of authors and contributors (alphabetic order)

Thomas S. Dye	University of Hawai'i at Manoa, USA

Nicolas Frerebeau	Université Bordeaux Montaigne, France

Anne Philippe	Université de Nantes, France

Marie-Anne Vibet	Université de Nantes, France

Package maintainer

Anne Philippe
anne.philippe@univ-nantes.fr

Laboratoire de Mathématiques Jean Leray (UMR 6629)
2, rue de la Houssinière
BP 92208
F-44322 Nantes Cedex 3
France