ArchaeoPhases: Post-Processing of the Markov Chain Simulated by 'ChronoModel', 'Oxcal' or 'BCal'

Provides a list of functions for the statistical analysis of archaeological dates and groups of dates. It is based on the post-processing of the Markov Chains whose stationary distribution is the posterior distribution of a series of dates. Such output can be simulated by different applications such as 'ChronoModel' <https://chronomodel.com/>, 'Oxcal' <https://c14.arch.ox.ac.uk/oxcal.html> or 'BCal' <https://bcal.shef.ac.uk/>. The only requirement is to have a csv file containing a sample from the posterior distribution.

See also

Author

Full list of authors and contributors (alphabetic order)

Thomas S. Dye	University of Hawai'i at Manoa, USA

Nicolas Frerebeau	Université Bordeaux Montaigne, France

Anne Philippe	Université de Nantes, France

Marie-Anne Vibet	Université de Nantes, France

Package maintainer

Anne Philippe
anne.philippe@univ-nantes.fr

Laboratoire de Mathématiques Jean Leray (UMR 6629)
2, rue de la Houssinière
BP 92208
F-44322 Nantes Cedex 3
France