The Basic Allen Relation Set

The Basic Allen Relation Set

allen_relation_code(...)

allen_relation_string(...)

allen_relation_concurrent(...)

allen_relation_distinct(...)

Arguments

...

Currently not used.

Value

allen_relation_code() returns a character vector of one-letter codes for the thirteen basic Allen relations.

allen_relation_string() returns a character vector of string descriptors of the Allen basic relations.

allen_relation_concurrent() returns a character vector of nine one-letter codes for the Allen concurrent relations.

allen_relation_distinct() returns the six value Allen relation set for intervals with distinct endpoints.

Note

The codes were proposed by Thomas Alspaugh.

References

Allen, J. F. (1983). Maintaining Knowledge about Temporal Intervals. Communications of the ACM, 26(11): 832-843. doi:10.1145/182.358434 .

Alspaugh, T. (2019). Allen's Interval Algebra. URL: https://thomasalspaugh.org/pub/fnd/allen.html.

See also

Author

T. S. Dye