Gap or Hiatus Between Successive Phases

Finds, if it exists, a gap or hiatus between two successive phases. This gap or hiatus is the longest interval that satisfies \(P(Phase1Max < IntervalInf < IntervalSup < Phase2Min | M) = level\)

hiatus(x, y, ...)

# S4 method for numeric,numeric
hiatus(x, y, level = 0.95)

# S4 method for PhasesMCMC,missing
hiatus(x, level = 0.95)

Arguments

x, y
A numeric vector. If y is missing, x must be an PhasesMCMC object.

...
Currently not used.

level
A length-one numeric vector giving the confidence level.

Value

A data.frame containing the endpoints (in years BC/AD) of the hiatus between each pair of successive phases (at a given level).

See also

Author

A. Philippe, M.-A. Vibet, N. Frerebeau

Examples

## Coerce to MCMC
events <- as_mcmc(Events, iteration = 1)
events <- events[1:10000, ]

## Compute min-max range for all chains
all <- phase(events)
head(all)

#> An object of class "PhasesMCMC"
#>   phase_1_start phase_1_end
#> 1      -1711.42    -732.129
#> 2      -1723.80    -739.132
#> 3      -1655.08    -633.252
#> 4      -1667.91    -699.241
#> 5      -1933.15    -821.100
#> 6      -1854.19    -719.353
#> Slot "start":
#> [1] 1
#> 
#> Slot "end":
#> [1] 2
#> 
#> Slot "ordered":
#> [1] FALSE
#> 
#> Slot "phases":
#> [1] "phase_1"
#> 
#> Slot "calendar":
#> [1] "BCAD"
#> 
#> Slot "hash":
#> character(0)
#> 

## Compute min-max range by group
phases <- phase(events, groups = list(A = c(1, 3), B = c(2, 4)))
head(phases)

#> An object of class "PhasesMCMC"
#>    A_start    A_end  B_start    B_end
#> 1 -872.515 -732.129 -1711.42 -1093.75
#> 2 -748.011 -739.132 -1723.80 -1192.21
#> 3 -917.723 -633.252 -1655.08 -1180.27
#> 4 -872.857 -699.241 -1667.91 -1226.39
#> 5 -884.704 -821.100 -1933.15 -1307.96
#> 6 -972.114 -719.353 -1854.19 -1217.13
#> Slot "start":
#> [1] 1 3
#> 
#> Slot "end":
#> [1] 2 4
#> 
#> Slot "ordered":
#> [1] FALSE
#> 
#> Slot "phases":
#> [1] "A" "B"
#> 
#> Slot "calendar":
#> [1] "BCAD"
#> 
#> Slot "hash":
#> character(0)
#> 

## Unless otherwise specified, the phases are assumed to be unordered
get_order(phases)

#> [1] A B
#> Levels: A B

## Set chronological order
## (from the older to the youngest phase)
set_order(phases) <- c("B", "A")
get_order(phases)

#> [1] B A
#> Levels: B < A

## Compute phase ranges
boundaries(phases)

#>   start   end
#> B -1972 -1078
#> A -1046  -202

## Compute phase transition
transition(phases)

#>               start  end
#> B_end-A_start -1409 -502

## Compute phase hiatus
hiatus(phases)

#>               start   end
#> B_end-A_start -1063 -1046