Bayesian Test for Anteriority/Posteriority Between Two Parameters

This function estimates the posterior probability that event x is older than event b using the output of the MCMC algorithm. This provides a Bayesian test for checking the following assumption: "Event x is older than event y".

test_older(x, y, ...)

# S4 method for numeric,numeric
test_older(x, y, ...)

Arguments

x
A numeric vector giving the output of the MCMC algorithm for the first parameter.

y
A numeric vector giving the output of the MCMC algorithm for the second parameter.

...
Currently not used.

Value

A numeric vector (the posterior probability of the assumption: "event x is older than event y").

Details

For a given output of MCMC algorithm, this function estimates the posterior probability of the event \(x < y\) by the relative frequency of the event "the value of event x is less than the value of event y" in the simulated Markov chain.

Author

A. Philippe, M.-A. Vibet, N. Frerebeau